Fixed some compilation problems

+++ projects/test_dex_analysis.cpp
@@ -51,7 +51,7 @@ int main(int argc, char **argv)
     auto lapse = std::chrono::duration_cast<std::chrono::microseconds>(end - start);
     
     //time in microseconds
-    std::cout << (double)(lapse.count()/1000000) << "\n";
+    std::cout << (static_cast<double>(lapse.count())/static_cast<double>(1000000)) << "\n";
 
     return 0;
 }
